Changed how spatial hash works
Spatial hash and spatial hash entry are now template classes so you can
use an ytype of object for them.

spatial hash entry now keeps a shared ptr to object instead of weak ptr.
Instead of keeping two seperate containers of an object and its spatial
hash entry counterpart it would be easier to just keep the spatial hash
entry, so now the entry uses a shared ptr so you won't need a seperate
object and an entry.

None of this is test yet.