Testing Hash, problem found
Tested the spatial hash to see if it offers any speed ups at all.

one problem with the hash is that it allows duplicates so an object
might end up checking for collision with another object several times.
When attempting to make the hash return a unique set of objects it
caused a significant slowdown so for now let the hash return duplicates.