SpawnPoints now specify what type of enemy to spawn
SpawnPoints now have a parameter that determines the type of enemy it
spawns. I decided to have each spawn point spawn a unique type of enemy
so if you want to spawn multiple types of enemies in one place you need
to create multiple spawn points, each assigned to a different enemy
type.