Testing implementation of parallax scrolling
tested a method of parallax scrolling by taking the views offset from
last frame to current frame and moving the background by some fraction
of that offset.

YOU CANNOT USE THE CAMERA OFFSET FOR THIS CALCULATION. the camera moves
even if the player is running into a wall, until the camera center hits
the edge of the screen. but since the camera center moves while the
player isn't moving then the parallax scrolling background will move
even though the player isn't moving. This problem doesn't occur if you
use the view center since the view center doesnt move if the player isnt
moving.