Added animation and state system to player
Players now draw an animated sprite to the screen instead of a white
box. Added hitbox state and animation states as well.

Bugs: overtime if player is standing on something it somehow switches to
the falling animation.

MAJOR ISSUE: hitboxes can't be different sizes otherwise when an object
switches to a different hitbox it might overlap with the ground or other
tiles and cause the object to be pushed out of the tile in some weird
direction.

perhaps there should be a central hitbox that handles collision with the
environment, and then seperate hurtboxes for each object that collides
with attacks and damages from other objects.