Added separate environment hitbox and entity hitbox
all entities have seperate hitboxes that collide with the environment,
and other entities. the environment hitbox,  collides with tiles and
dynamic objects, and generally doesn't change when the entity changes
state. the entity hitbox, called hurtbox, changes with every frame of an
animation so that its always accurate to the entity.