Changed PreloadedDataCollection class
PreloadedDataCollection class now keeps a map object for every type of
object whose data can be cached. The map key is an object's type, and
the map's value is the cache for that object type. accessing any data is
done using the public interface, which will return cached data if it's
available, or it will load data from a file and cache it before
returning it if the data hasn't already been loaded.