Fixed slope bugs
Bug fixed - when going down a slope that goes down to the right, the
player teleports to the top of the slope once he reaches the bottom.

this happened because when reaching the bottom the player is snapped to
the bottom of the slope, but the bottom of the slope was calculated by
adding the y-intercept of the slope to the yposition of the slope. for
slopes that go down to the right the y-intercept was 0 so players were
sent to the top of the slope.

the game now calculates the actual height of the bottom of the slope
using y=mx + b. this is required because there mgiht be half slopes so
just setting the players position to the bottom of the tile itself
wouldn't work.